requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
more-itertools==4.2.0
numpy==1.14.4
OpenCC==0.2
Pillow==5.1.0
pluggy==0.6.0
py==1.5.3
pytest==3.6.1
pytest-runner==4.2
ruamel.yaml==0.15.37
six==1.11.0
